intersight-metrics-bridge/requirements.txt
influxdb-client
intersight
python-dotenv
PyYAML
requests
urllib3